extmod/modutimeq: Implement remove(item_handle) method.
item_handle is a return value from .push() method. The complexity of
remove() operation is the same as .pop() method, i.e. O(log N). Of
course, to achieve this complexity, it was required to add a "pos"
backreference to each queue entry, pointing to its current index in
the heap.

Together with converting a heap from heap of items to heap to pointers
to items, as done in a previous commit, to get a stable references to
items in the heap, this effectively means that there's an overhead of
2 pointers per item to maintain such a heap with "remove item" support.
Which makes the overhead as high as an efficient balanced binary tree
implementation (like red-black tree, as non-efficient binary tree
besides 2 child pointers would also require a parent pointer). Of course,
code size for a balanced binary tree implementation will be certainly
bigger than for a heap, but still, it's an interesting observation
that data storage space wise, we don't win with a heap under such
condition (efficient removal of an arbitrary item).